+++ packages/testland/src/constants.ts
@@ -24,6 +24,7 @@ export const AUTH_URL = env.AUTH_URL
 export const COUNTRY_CONFIG_URL = env.COUNTRY_CONFIG_URL
 
 export const SENTRY_DSN = env.SENTRY_DSN
+export const ENVIRONMENT_NAME = env.ENVIRONMENT_NAME
 
 export const PRODUCTION = env.isProd
 export const ESIGNET_REDIRECT_URL = env.ESIGNET_REDIRECT_URL